+++ Changelog.md
@@ -11,6 +11,7 @@ a) full typing; and b) ease of use; and c) consistency.  New parameterized is:
 first three params (`TChildProps` can be derived). [#1402](https://github.com/apollographql/react-apollo/pull/1402) 
 - Typescript - fix `graphql` HOC inference [#1402](https://github.com/apollographql/react-apollo/pull/1402)
 - **Remove deprecated** `operationOptions.options.skip`, use `operationOptions.skip` instead
+- Added <Query /> component [#1399](https://github.com/apollographql/react-apollo/pull/1398)
 
 ### 2.0.4
 - rolled back on the lodash-es changes from